user account rest flow

+++ src/controllers/user_controller.cr
@@ -1,11 +1,15 @@
 class UserController < ApplicationController
-  def index
-    users = User.all
-    render "index.slang"
+  authorize_with UserPolicy, User
+
+  private def user_params
+    params.to_h.select [
+      "email"
+    ]
   end
 
   def show
     if user = User.find params["id"]
+      authorize user
       render "show.slang"
     else
       flash["warning"] = "User with ID #{params["id"]} Not Found"
@@ -15,19 +19,20 @@ class UserController < ApplicationController
 
   def new
     user = User.new
-    password = ""
+    authorize user
     render "new.slang"
   end
 
   def create
-    password = params["password"]
-    user = User.new params.to_h.select(["email"])
+    user = User.new user_params
+    user.hash_password params["password"]
 
-    user.save_password params["password"]
+    authorize user
 
     if user.valid? && user.save
-      flash["success"] = "Created User successfully."
-      redirect_to "/users"
+      flash["success"] = "Successfully registered and logged in."
+      login_user user
+      redirect_to "/"
     else
       flash["danger"] = "Could not create User!"
       render "new.slang"
@@ -35,8 +40,9 @@ class UserController < ApplicationController
   end
 
   def edit
-    password = ""
     user = User.find params["id"]
+
+    authorize user
 
     unless user
       flash["warning"] = "User with ID #{params["id"]} Not Found"
@@ -50,17 +56,18 @@ class UserController < ApplicationController
   def update
     user = User.find(params["id"])
 
+    authorize user
+
     unless user
       flash["warning"] = "User with ID #{params["id"]} Not Found"
       redirect_to "/users"
       return
     end
 
-    password = params["password"]
-    user.set_attributes params.to_h.select(["name", "email", "crypted_password"])
+    user.set_attributes user_params
 
-    if params["password"]
-      user.save_password params["password"]
+    if password = params["password"]
+      user.hash_password params["password"]
     end
 
     if user.valid? && user.save
@@ -72,12 +79,18 @@ class UserController < ApplicationController
     end
   end
 
+  def delete
+    render "delete.slang"
+  end
+
   def destroy
     if user = User.find params["id"]
+      authorize user
       user.destroy
     else
       flash["warning"] = "User with ID #{params["id"]} Not Found"
     end
+
     redirect_to "/users"
   end
 end
